requirements.txt
Flask==2.3.3
Flask-Limiter==3.5.0
Flask-WTF==1.1.1
WTForms==3.0.1
gunicorn==21.2.0
google-generativeai==0.3.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.7
